deplacement findBooking dans le store
